Move Timeline to the data package
