+++ service/src/main/kotlin/fi/espoo/evaka/reservations/ReservationQueries.kt
@@ -16,7 +16,7 @@ import fi.espoo.evaka.shared.EvakaUserId
 import fi.espoo.evaka.shared.HolidayQuestionnaireId
 import fi.espoo.evaka.shared.PersonId
 import fi.espoo.evaka.shared.PlacementId
-import fi.espoo.evaka.shared.Timeline
+import fi.espoo.evaka.shared.data.Timeline
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
 import fi.espoo.evaka.shared.domain.DateRange
